additional file restructuring
