+++ Frontend/src/components/mentee/mentee-dashboard/MenteeDashboard.jsx
@@ -1,7 +1,7 @@
 import { useState} from 'react'
 import './MenteeDashboard.css'
 import Calendar from '../../Calendar'
-import Requests from '../../requests/Requests'
+import Requests from '../../mentor/mentor-requests/Requests'
 import { Link } from 'react-router-dom';
 import MenteeMatches from '../mentee-matches/MenteeMatches'
