+++ Frontend/src/components/mentee/mentee-matches/MenteeMatches.jsx
@@ -1,8 +1,8 @@
 import { useContext, useEffect, useState } from 'react';
 import { UserContext } from '../../../UserContext.jsx';
-import Request from '../../requests/Request.jsx';
+import Request from '../../mentor/mentor-requests/Request.jsx';
 import MatchCard from '../../mentor-matching/MatchCard.jsx';
-import './MentorMatches.css'
+import '../../mentor/mentor-matches/MentorMatches.css'
 import config  from '../../../../config.js';
 
 function MenteeMatches() {
