+++ Frontend/src/components/mentor-matching/MatchCard.jsx
@@ -1,7 +1,7 @@
 import { useState} from 'react'
 import { UserContext } from '../../UserContext.jsx';
 import config from '../../../config.js';
-import "./requests/Request.css"
+import "../mentor/mentor-requests/Request.css"
 
 const PLACEHOLDER = "https://ralfvanveen.com/wp-content/uploads/2021/06/Placeholder-_-Glossary.svg";
